# Replace mapping.emplace(K, V) with call to mapping.find(K)